+++ tools/db_crashtest.py
@@ -279,7 +279,6 @@ blob_params = {
     "blob_garbage_collection_age_cutoff": lambda: random.choice([0.0, 0.25, 0.5, 0.75, 1.0]),
     # The following are currently incompatible with the integrated BlobDB
     "use_merge": 0,
-    "backup_one_in": 0,
 }
 
 ts_params = {
@@ -299,7 +298,6 @@ ts_params = {
     "use_blob_db": 0,
     "enable_compaction_filter": 0,
     "ingest_external_file_one_in": 0,
-    "partition_filters": 0,
 }
 
 def finalize_and_sanitize(src_params):